+++ packages/hardhat-chai-matchers/test/fixture-projects/hardhat-project/contracts/Events.sol
@@ -9,6 +9,7 @@ contract Events {
   event WithoutArgs();
   event WithUintArg(uint u);
   event WithIntArg(int i);
+  event WithAddressArg(address a);
   event WithTwoUintArgs(uint u, uint v);
   event WithStringArg(string s);
   event WithTwoStringArgs(string s, string t);
@@ -37,6 +38,10 @@ contract Events {
 
   function emitInt(int i) public {
     emit WithIntArg(i);
+  }
+
+  function emitAddress(address a) public {
+    emit WithAddressArg(a);
   }
 
   function emitUintTwice(uint u, uint v) public {