+++ packages/hardhat-toolbox-viem/CHANGELOG.md
@@ -1,1 +1,7 @@
 # @nomicfoundation/hardhat-toolbox-viem
+
+## 2.0.0
+
+### Major Changes
+
+- 23665f399: Upgraded hardhat-verify dependency